Module log4rs::append::rolling_file

A rolling file appender.

Logging directly to a file can be a dangerous proposition for long running processes. You wouldn’t want to start a server up and find out a couple weeks later that the disk is filled with hundreds of gigabytes of logs! A rolling file appender alleviates these issues by limiting the amount of log data that’s preserved.

Like a normal file appender, a rolling file appender is configured with the location of its log file and the encoder which formats log events written to it. In addition, it holds a “policy” object which controls when a log file is rolled over and how the old files are archived.

For example, you may configure an appender to roll the log over once it reaches 50 megabytes, and to preserve the last 10 log files.

Requires the rolling_file_appender feature.

Modules

policy

Policies.

Structs

LogFile

Information about the active log file.

RollingFileAppender

An appender which archives log files in a configurable strategy.

RollingFileAppenderBuilder

A builder for the RollingFileAppender.

RollingFileAppenderConfig

Configuration for the rolling file appender.

RollingFileAppenderDeserializer

A deserializer for the RollingFileAppender.