Struct log4rs::append::rolling_file::policy::compound::roll::fixed_window::FixedWindowRollerBuilder

pub struct FixedWindowRollerBuilder { /* fields omitted */ }

A builder for the FixedWindowRoller.

Methods

impl FixedWindowRollerBuilder

fn base(self, base: u32) -> FixedWindowRollerBuilder

Sets the base index for archived log files.

Defaults to 0.

fn build(self,
pattern: &str,
count: u32)
-> Result<FixedWindowRoller, Box<Error + Sync + Send>>

Constructs a new FixedWindowRoller.

pattern must contain at least one instance of {}, all of which will be replaced with an archived log file's index.

If the file extension of the pattern is .gz and the gzip Cargo feature is enabled, the archive files will be gzip-compressed.