Struct log4rs::append::rolling_file::policy::compound::trigger::size::SizeTrigger

pub struct SizeTrigger { /* fields omitted */ }

A trigger which rolls the log once it has passed a certain size.

Methods

impl SizeTrigger

fn new(limit: u64) -> SizeTrigger

Returns a new trigger which rolls the log once it has passed the specified size in bytes.