Trait log4rs::append::rolling_file::policy::Policy

pub trait Policy: Sync + Send + 'static + Debug {
    fn process(
        &self, 
        log: &mut LogFile<'_>
    ) -> Result<(), Box<dyn Error + Sync + Send>>;
}

A trait implementing a rolling policy for a RollingFileAppender.

Required methods

fn process(
&self,
log: &mut LogFile<'_>
) -> Result<(), Box<dyn Error + Sync + Send>>

Rolls the current log file, if necessary.

This method is called after each log event. It is provided a reference to the current log file.