log4rs::config
Function init_file

pub fn init_file<P>(path: P, deserializers: Deserializers) -> Result<()>
where
    P: AsRef<Path>,

Initializes the global logger as a log4rs logger configured via a file.

Configuration is read from a file located at the provided path on the filesystem and components are created from the provided Deserializers.

Any nonfatal errors encountered when processing the configuration are reported to stderr.

Requires the file feature (enabled by default).