Struct log4rs::config::runtime::Logger

pub struct Logger { /* private fields */ }

Configuration for a logger.

Implementations

impl Logger

pub fn builder() -> LoggerBuilder

Creates a new LoggerBuilder with the specified name and level.

There are initially no appenders attached and additive is true.

pub fn name(&self) -> &str

Returns the name of the logger.

pub fn level(&self) -> LevelFilter

Returns the minimum level of log messages that the logger will accept.

pub fn appenders(&self) -> &[String]

Returns the list of names of appenders that will be attached to the logger.

pub fn additive(&self) -> bool

Determines if appenders of parent loggers will also be attached to this logger.