Struct log4rs::config::ConfigBuilder

pub struct ConfigBuilder { /* fields omitted */ }

A builder for Configs.

Methods

impl ConfigBuilder

fn appender(self, appender: Appender) -> ConfigBuilder

Adds an appender.

fn appenders<I>(self, appenders: I) -> ConfigBuilder where I: IntoIterator<Item=Appender>

Adds appenders.

fn logger(self, logger: Logger) -> ConfigBuilder

Adds a logger.

fn loggers<I>(self, loggers: I) -> ConfigBuilder where I: IntoIterator<Item=Logger>

Adds loggers.

fn build_lossy(self, root: Root) -> (Config, Vec<Error>)

Consumes the ConfigBuilder, returning the Config.

Unlike build, this method will always return a Config by stripping portions of the configuration that are incorrect.

fn build(self, root: Root) -> Result<Config, Errors>

Consumes the ConfigBuilder, returning the Config.