Struct log4rs::config::Deserializers

pub struct Deserializers(_);

A container of Deserializers.

Implementations

impl Deserializers

pub fn new() -> Deserializers

Creates a Deserializers with default mappings.

All are enabled by default.

Appenders

“console” -> ConsoleAppenderDeserializer

Requires the console_appender feature.

“file” -> FileAppenderDeserializer

Requires the file_appender feature.

“rolling_file” -> RollingFileAppenderDeserializer

Requires the rolling_file_appender feature.

Encoders

“pattern” -> PatternEncoderDeserializer

Requires the pattern_encoder feature.

“json” -> JsonEncoderDeserializer

Requires the json_encoder feature.

Filters

“threshold” -> ThresholdFilterDeserializer

Requires the threshold_filter feature.

Policies

“compound” -> CompoundPolicyDeserializer

Requires the compound_policy feature.

Rollers

“delete” -> DeleteRollerDeserializer

Requires the delete_roller feature.

“fixed_window” -> FixedWindowRollerDeserializer

Requires the fixed_window_roller feature.

Triggers

“size” -> SizeTriggerDeserializer

Requires the size_trigger feature.

pub fn empty() -> Deserializers

Creates a new Deserializers with no mappings.

pub fn insert<T>(&mut self, kind: &str, deserializer: T)where T: Deserialize,

Adds a mapping from the specified kind to a deserializer.

pub fn deserialize<T>(&self, kind: &str, config: Value) -> Result<Box<T>>where T: Deserializable + ?Sized,

Deserializes a value of a specific type and kind.