Struct log4rs::config::Logger

pub struct Logger {
    // some fields omitted
}

Configuration for a logger.

Methods

impl Logger

fn builder(name: String, level: LogLevelFilter) -> LoggerBuilder

Creates a new LoggerBuilder with the specified name and level.

There are initially no appenders attached and additive is true.

fn name(&self) -> &str

Returns the name of the logger.

fn level(&self) -> LogLevelFilter

Returns the minimum level of log messages that the logger will accept.

fn appenders(&self) -> &[String]

Returns the list of names of appenders that will be attached to the logger.

fn additive(&self) -> bool

Determines if appenders of parent loggers will also be attached to this logger.