Module log4rs::encode

Encoders

Modules

json

An encoder which writes a JSON object.

pattern

A simple pattern-based encoder.

writer

Implementations of the encode::Write trait.

Structs

EncoderConfig

Configuration for an encoder.

Style

The style applied to text output.

Enums

Color

A text or background color.

Traits

Encode

A trait implemented by types that can serialize a Record into a Writer.

Write

A trait for types that an Encoder will write to.