Struct log4rs::encode::writer::ConsoleWriter

pub struct ConsoleWriter(_);

An encode::Writer that outputs to a console.

Methods

impl ConsoleWriter

fn stdout() -> Option<ConsoleWriter>

Returns a new ConsoleWriter that will write to standard out.

Returns None if standard out is not a console buffer on Windows, and if it is not a tty on Unix.

fn lock<'a>(&'a self) -> ConsoleWriterLock<'a>

Locks the console, preventing other threads from writing concurrently.