Enum log4rs::FilterResponse

pub enum FilterResponse {
    Accept,
    Neutral,
    Reject,
}

The response returned by a filter.

Variants

Accept

Accept the log event.

It will be immediately passed to the appender, bypassing any remaining filters.

Neutral

Take no action on the log event.

It will continue on to remaining filters or pass on to the appender if there are none remaining.

Reject

Reject the log event.