Struct log4rs::file::Config

pub struct Config {
    // some fields omitted
}

A deserialized log4rs configuration file.

Methods

impl Config

fn parse(config: &str, format: Format, deserializers: &Deserializers) -> Result<Config, Box<Error>>

Creates a log4rs Config from the specified config string and Deserializers.

fn refresh_rate(&self) -> Option<Duration>

Returns the requested refresh rate.

fn into_config(self) -> Config

Returns the log4rs Config.

fn errors(&self) -> &[Error]

Returns any nonfatal errors encountered when deserializing the config.