Struct log4rs::file::Deserializers

pub struct Deserializers(_);

A container of Deserializers.

Methods

impl Deserializers

fn new() -> Deserializers

Creates a new Deserializers with no mappings.

fn insert<T>(&mut self, kind: &str, deserializer: T) where T: Deserialize

Adds a mapping from the specified kind to a deserializer.

fn deserialize<T: ?Sized>(&self, kind: &str, config: Value) -> Result<Box<T>, Box<Error>> where T: Deserializable

Deserializes a value of a specific type and kind.