Trait log4rs::filter::Filter

pub trait Filter: Debug + Send + Sync + 'static {
    fn filter(&self, record: &Record<'_>) -> Response;
}

The trait implemented by log4rs filters.

Filters are associated with appenders and limit the log events that will be sent to that appender.

Required Methods

fn filter(&self, record: &Record<'_>) -> Response

Filters a log event.