Function log4rs::init

pub fn init(config_file: Path, loggers: HashMap<String, Box<MakeLogger>>) -> Result<(), SetLoggerError>

Initializes the global logger with a log4rs logger.

Configuration is read from a TOML file located at the provided path on the filesystem, and log4rs loggers are created from the provided set of logger makers.

Any errors encountered when processing the configuration are reported to stderr.