Function log4rs::init_file

pub fn init_file<P: AsPath + ?Sized>(path: &P, creator: Creator) -> Result<(), SetLoggerError>

Initializes the global logger with a log4rs logger.

Configuration is read from a TOML file located at the provided path on the filesystem and appenders are created from the provided Creator.

Any errors encountered when processing the configuration are reported to stderr.