Struct log4rs::loggers::FileLoggerMaker

pub struct FileLoggerMaker;

A logger maker returning a log4rs logger which logs to a specified file.

The file is specified by the file parameter. The log output format may be customized via the pattern parameter. The accepted syntax is documented in the pattern module.

The default_loggers function maps the string "file" to this logger maker.

Examples

[[logger.errors]]
kind = "file"
file = "log/errors.log"
pattern = "%d [%l] %M - %m"