Struct log4rs::pattern::PatternLayout

pub struct PatternLayout {
    // some fields omitted
}

A formatter object for LogRecords.

Methods

impl PatternLayout

fn new(pattern: &str) -> Result<PatternLayout, Error>

Creates a PatternLayout from a pattern string.

The pattern string syntax is documented in the pattern module.

fn append<W>(&self, w: &mut W, record: &LogRecord) -> Result<()> where W: Write

Writes the specified LogRecord to the specified Writer according to its pattern.