Struct log4rs::toml::Creator

pub struct Creator {
    // some fields omitted
}

A type that can create appenders.

Creator implementes Default, which returns a Creator with mappings from "file" to FileAppenderCreator and "console" to ConsoleAppenderCreator.

Methods

impl Creator

fn new() -> Creator

Creates a new Creator with no appender mappings.

fn add_appender(&mut self, kind: &str, creator: Box<CreateAppender>)

Adds a mapping from the specified kind to the specified appender creator.