Struct log4rs::toml::FileAppenderCreator

pub struct FileAppenderCreator;

An appender creator for the FileAppender.

The path key is required, and specifies the path to the log file. The pattern key is optional and specifies a PatternLayout pattern to be used for output. The append key is optional and specifies whether the output file should be truncated or appended to.