Crate log4rs

log4rs is a highly configurable logging framework for Rust integrating with the log logging facade.

The log4rs framework consists of a number of loggers as well as directies which configure which log messages are directed to whic loggers. Configuration is handled via a TOML file:

# log4rs will automatically check the configuration file for updates at the
# specified frequency in seconds if this is set
refresh_rate = 30

# A minimal configuration of a logger called "stderr" of the "console" kind
[logger.stderr]
kind = "console"

# A logger called "errors" with a more advanced configuration.
[logger.errors]
kind = "file"
path = "logs/errors.log"
pattern = "%d [%l] %T - %m"

# A directive that routes all error level log messages to the errors logger
[[directive]]
level = "error"
logger = "errors"

# A directive that routes all error, warning, and info level messages from
# the foo::bar::baz module to the stderr logger.
[[directive]]
level = "info"
logger = "stderr"
path = "foo::bar::baz"

The init method should be called early in the runtime of a program to initialize the logging system. It takes a path to the config file as well as a map of available loggers. The default_loggers function returns a map of loggers that are provided with log4rs, including the console and file kinds mentioned above:

extern crate log4rs;

fn main() {
    log4rs::init(Path::new("log.toml"), log4rs::default_loggers()).unwrap();

    // ...
}

Modules

loggers
A set of basic log4rs loggers.

pattern
Simple pattern syntax for logger output formats.

Traits

Log
A trait specifying the interface for log4rs loggers.

MakeLogger
A trait specifying the interface for creation of log4rs loggers.

Functions

default_loggers
Returns a set of built-in log4rs loggers.

init
Initializes the global logger with a log4rs logger.