log4rs::append::rolling_file::policy::compound::roll
Trait Roll

pub trait Roll:
    Debug
    + Send
    + Sync
    + 'static {
    // Required method
    fn roll(&self, file: &Path) -> Result<()>;
}

A trait which processes log files after they have been rolled over.

Required Methods

fn roll(&self, file: &Path) -> Result<()>

Processes the log file.

At the time that this method has been called, the log file has already been closed.

If this method returns successfully, there must no longer be a file at the specified location.