Trait log4rs::append::rolling_file::policy::compound::roll::Roll

pub trait Roll: Debug + Send + Sync + 'static {
    fn roll(&self, file: &Path) -> Result<(), Box<dyn Error + Sync + Send>>;
}

A trait which processes log files after they have been rolled over.

Required methods

fn roll(&self, file: &Path) -> Result<(), Box<dyn Error + Sync + Send>>

Processes the log file.

At the time that this method has been called, the log file has already been closed.

If this method returns successfully, there must no longer be a file at the specified location.

Loading content...