Trait log4rs::append::rolling_file::policy::compound::trigger::Trigger

pub trait Trigger: Debug + Send + Sync + 'static {
    fn trigger(&self, file: &LogFile) -> Result<bool, Box<Error>>;
}

A trait which identifies if the active log file should be rolled over.

Required Methods

fn trigger(&self, file: &LogFile) -> Result<bool, Box<Error>>

Determines if the active log file should be rolled over.