log4rs::append::rolling_file::policy
Trait Policy

pub trait Policy:
    Sync
    + Send
    + 'static
    + Debug {
    // Required methods
    fn process(&self, log: &mut LogFile<'_>) -> Result<()>;
    fn is_pre_process(&self) -> bool;
}

A trait implementing a rolling policy for a RollingFileAppender.

Required Methods

fn process(&self, log: &mut LogFile<'_>) -> Result<()>

Rolls the current log file, if necessary.

This method is called after each log event. It is provided a reference to the current log file.

fn is_pre_process(&self) -> bool

Return the config Trigger.is_pre_process value