Trait log4rs::append::rolling_file::policy::Policy

pub trait Policy: Sync + Send + 'static + Debug {
    fn process(&self, log: &mut LogFile<'_>) -> Result<()>;
}

A trait implementing a rolling policy for a RollingFileAppender.

Required Methods

fn process(&self, log: &mut LogFile<'_>) -> Result<()>

Rolls the current log file, if necessary.

This method is called after each log event. It is provided a reference to the current log file.