log4rs::append
Trait Append

pub trait Append:
    Debug
    + Send
    + Sync
    + 'static {
    // Required methods
    fn append(&self, record: &Record<'_>) -> Result<()>;
    fn flush(&self);
}

A trait implemented by log4rs appenders.

Appenders take a log record and processes them, for example, by writing it to a file or the console.

Required Methods

fn append(&self, record: &Record<'_>) -> Result<()>

Processes the provided Record.

fn flush(&self)

Flushes all in-flight records.