Trait log4rs::append::Append

pub trait Append: Debug + Send + Sync + 'static {
    fn append(
        &self, 
        record: &Record
    ) -> Result<(), Box<dyn Error + Sync + Send>>;

    fn flush(&self);
}

A trait implemented by log4rs appenders.

Appenders take a log record and processes them, for example, by writing it to a file or the console.

Required methods

fn append(&self, record: &Record) -> Result<(), Box<dyn Error + Sync + Send>>

Processes the provided Record.

fn flush(&self)

Flushes all in-flight records.

Loading content...