Struct log4rs::config::ConfigBuilder

pub struct ConfigBuilder {
    // some fields omitted
}

A builder for Configs.

Methods

impl ConfigBuilder

fn appender(self, appender: Appender) -> ConfigBuilder

Adds an appender.

fn appenders<I: IntoIterator<Item=Appender>>(self, appenders: I) -> ConfigBuilder

Adds appenders.

fn logger(self, logger: Logger) -> ConfigBuilder

Adds a logger.

fn loggers<I: IntoIterator<Item=Logger>>(self, loggers: I) -> ConfigBuilder

Adds loggers.

fn build_lossy(self, root: Root) -> (Config, Result<(), Errors>)

Consumes the ConfigBuilder, returning the Config.

Unlike build, this method will always return a Config by stripping portions of the configuration that are incorrect.

fn build(self, root: Root) -> Result<Config, Errors>

Consumes the ConfigBuilder, returning the Config.