Trait log4rs::config::Deserializable

pub trait Deserializable: 'static {
    fn name() -> &'static str;
}

A trait implemented by traits which are deserializable.

Required Methods

fn name() -> &'static str

Returns a name for objects implementing the trait suitable for display in error messages.

For example, the Deserializable implementation for the Append trait returns “appender”.