Struct log4rs::encode::Style

pub struct Style {
    pub text: Option<Color>,
    pub background: Option<Color>,
    pub intense: Option<bool>,
    // some fields omitted
}

The style applied to text output.

Any fields set to None will be set to their default format, as defined by the Writer.

Fields
text: Option<Color>

The text (or foreground) color.

background: Option<Color>

The background color.

intense: Option<bool>

True if the text should have increased intensity.

Methods

impl Style

fn new() -> Style

Returns a Style with all fields set to their defaults.

fn text(&mut self, text: Color) -> &mut Style

Sets the text color.

fn background(&mut self, background: Color) -> &mut Style

Sets the background color.

fn intense(&mut self, intense: bool) -> &mut Style

Sets the text intensity.