log4rs::encode
Trait Encode

pub trait Encode:
    Debug
    + Send
    + Sync
    + 'static {
    // Required method
    fn encode(&self, w: &mut dyn Write, record: &Record<'_>) -> Result<()>;
}

A trait implemented by types that can serialize a Record into a Writer.

Encoders are commonly used by Appenders to format a log record for output.

Required Methods

fn encode(&self, w: &mut dyn Write, record: &Record<'_>) -> Result<()>

Encodes the Record into bytes and writes them.