log4rs::encode
Trait Write

pub trait Write: Write {
    // Provided method
    fn set_style(&mut self, style: &Style) -> Result<()> { ... }
}

A trait for types that an Encoder will write to.

It extends std::io::Write and adds some extra functionality.

Provided Methods

fn set_style(&mut self, style: &Style) -> Result<()>

Sets the output text style, if supported.

Writers should ignore any parts of the Style they do not support.

The default implementation returns Ok(()). Implementations that do not support styling should do this as well.