Struct log4rs::toml::Creator

pub struct Creator {
    // some fields omitted
}

A type that can create appenders.

Creator implements Default, which returns a Creator with the following mappings:

Appenders

"file" -> FileAppenderCreator

"console" -> ConsoleAppenderCreator

Filters

"threshold" -> ThresholdFilterCreator

Methods

impl Creator

fn new() -> Creator

Creates a new Creator with no appender or filter mappings.

fn add_appender(&mut self, kind: &str, creator: Box<CreateAppender>)

Adds a mapping from the specified kind to the specified appender creator.

fn add_filter(&mut self, kind: &str, creator: Box<CreateFilter>)

Adds a mapping from the specified kind to the specified filter creator.